requirements.txt
streamlit
langchain
langchain-openai
chromadb
pypdf
python-dotenv
tiktoken
langchain-community